refactor: split PNG fallback resizer facade
