+++ src/hooks/read-image-resizer/png-constants.ts
@@ -1,0 +1,1 @@
+export const PNG_SIGNATURE = Buffer.from([0x89, 0x50, 0x4e, 0x47, 0x0d, 0x0a, 0x1a, 0x0a])
