+++ src/hooks/read-image-resizer/png-crc.ts
@@ -1,0 +1,31 @@
+const CRC_TABLE = buildCrcTable()
+
+function buildCrcTable(): Uint32Array {
+  const table = new Uint32Array(256)
+
+  for (let n = 0; n < 256; n++) {
+    let c = n
+
+    for (let k = 0; k < 8; k++) {
+      if (c & 1) {
+        c = 0xedb88320 ^ (c >>> 1)
+      } else {
+        c = c >>> 1
+      }
+    }
+
+    table[n] = c
+  }
+
+  return table
+}
+
+export function crc32(data: Buffer): number {
+  let crc = 0xffffffff
+
+  for (let i = 0; i < data.length; i++) {
+    crc = CRC_TABLE[(crc ^ data[i]) & 0xff] ^ (crc >>> 8)
+  }
+
+  return (crc ^ 0xffffffff) >>> 0
+}
